Added Hall Effect sensor, added LCD display methods, updated protocols, and a few additional changes.

+++ SecuritySystem/Structures/SensorData.h
@@ -43,7 +43,7 @@ typedef struct {
    uint16_t Temperature;
    uint16_t TempDecimal;
    uint16_t Proximity[NUM_PROXIMITY_SENSORS];
-   uint16_t HallEffect[NUM_HALLEFFECT_SENSORS];
+   uint8_t HallEffect[NUM_HALLEFFECT_SENSORS];
 
    uint16_t RTC[19];
    DateTime DateTime;
